New ScrapeMetadata feature, make scraping optional.

+++ root/scripts/script.sh
@@ -432,20 +432,23 @@ for folder in $(ls /input); do
 	# remove empty directories
 	find /input/$folder -mindepth 1 -type d -empty -exec rm -rf {} \; &>/dev/null
 	
-	# Scrape from screenscraper
-	Skyscraper -f emulationstation -u $ScreenscraperUsername:$ScreenscraperPassword -p $folder -d /cache/$folder -s screenscraper -i /output/$folder --flags relative,videos,unattend,nobrackets,unpack
-	# Save scraped data to output folder
-	Skyscraper -f emulationstation -p $folder -d /cache/$folder -i /output/$folder --flags relative,videos,unattend,nobrackets
-	# Remove skipped roms
-	if [ -f /root/.skyscraper/skipped-$folder-cache.txt ]; then
-		cat /root/.skyscraper/skipped-$folder-cache.txt | while read LINE;
-		do 
-			rm "$LINE"
-		done
-	fi	
+	if [ "$ScrapeMetadata" = "true" ]; then
+		echo "$ConsoleName :: Begin Skyscraper Process..."
+		# Scrape from screenscraper
+		Skyscraper -f emulationstation -u $ScreenscraperUsername:$ScreenscraperPassword -p $folder -d /cache/$folder -s screenscraper -i /output/$folder --flags relative,videos,unattend,nobrackets,unpack
+		# Save scraped data to output folder
+		Skyscraper -f emulationstation -p $folder -d /cache/$folder -i /output/$folder --flags relative,videos,unattend,nobrackets
+		# Remove skipped roms
+		if [ -f /root/.skyscraper/skipped-$folder-cache.txt ]; then
+			cat /root/.skyscraper/skipped-$folder-cache.txt | while read LINE;
+			do 
+				rm "$LINE"
+			done
+		fi
+	else
+		echo "$ConsoleName :: Metadata Scraping disabled..."
+		echo "$ConsoleName :: Enable by setting \"ScrapeMetadata=true\""
+	fi
 	
-	# set permissions
-	find /output/$folder -type d -exec chmod 777 {} \;
-	find /output/$folder -type f -exec chmod 666 {} \;
 done
 exit $?
